@@ -13,7 +13,7 @@
 		"@sveltejs/kit": "^1.5.0",
 		"autoprefixer": "^10.4.14",
 		"postcss": "^8.4.23",
-		"svelte": "^3.54.0",
+		"svelte": "^3.59.1",
 		"tailwindcss": "^3.3.2",
 		"vite": "^4.3.0"
 	},
@@ -35,6 +35,7 @@
 		"flowbite": "^1.6.5",
 		"flowbite-svelte": "^0.35.3",
 		"preline": "^1.8.0",
+		"preprocess": "^3.2.0",
 		"uuid": "^9.0.0"
 	}
 }
